+++ example/lib/medium_clap.dart
@@ -43,9 +43,21 @@ class _ClapButtonState extends State<ClapButton> {
       // aligns the top-center of `child` with the bottom-center of `portal`
       childAnchor: Alignment.topCenter,
       portalAnchor: Alignment.bottomCenter,
-      portal: Material(
-        elevation: 8,
-        borderRadius: BorderRadius.circular(40),
+      closeDuration: kThemeChangeDuration,
+      portal: TweenAnimationBuilder<double>(
+        tween: Tween(begin: 0, end: hasClappedRecently ? 1 : 0),
+        duration: kThemeChangeDuration,
+        builder: (context, progress, child) {
+          return Material(
+            elevation: 8 * progress,
+            animationDuration: Duration.zero,
+            borderRadius: const BorderRadius.all(Radius.circular(40)),
+            child: Opacity(
+              opacity: progress,
+              child: child,
+            ),
+          );
+        },
         child: Padding(
           padding: const EdgeInsets.all(8),
           child: Text('$clapCount'),